+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/apis/HubdleEnableableExtension.kt
@@ -61,7 +61,7 @@ public abstract class HubdleEnableableExtension(
         }
 
     internal fun applicablePlugin(
-        isEnabled: Property<Boolean>,
+        isEnabled: Provider<Boolean>,
         priority: Priority,
         scope: ApplicablePlugin.Scope,
         pluginId: PluginId,