+++ hubdle-gradle-plugin/testFunctional/resources/kotlin/jvm/features/molecule/main/kotlin/com/javiersc/kotlin/jvm/molecule/Main.kt
@@ -1,6 +1,6 @@
 package com.javiersc.kotlin.jvm.molecule
 
-import app.cash.molecule.RecompositionClock
+import app.cash.molecule.RecompositionMode
 import app.cash.molecule.moleculeFlow
 import kotlinx.coroutines.runBlocking
 
@@ -8,12 +8,13 @@ fun main() {
     runCatching {
         runBlocking {
             val store = FooMoleculeStore(FooDatasource())
-            moleculeFlow(RecompositionClock.Immediate) { store.state() }.collect {
-                if (it.counter == 1) store.effect(FooEffect.A)
-                if (it.counter == 3) store.effect(FooEffect.B)
-                if (it.counter == 5) error(it.counter)
-                println(it)
-            }
+            moleculeFlow(RecompositionMode.Immediate) { store.state() }
+                .collect {
+                    if (it.counter == 1) store.effect(FooEffect.A)
+                    if (it.counter == 3) store.effect(FooEffect.B)
+                    if (it.counter == 5) error(it.counter)
+                    println(it)
+                }
         }
     }
 }