+++ hubdle-version-catalog/CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `app.cash.molecule:app.cash.molecule.gradle.plugin -> 1.0.0`
+- `app.cash.molecule:app.cash.molecule.gradle.plugin -> 1.1.0`
 - `androidx.compose.compiler:compiler -> 1.5.0`
 - `org.jetbrains.kotlin.plugin.serialization:org.jetbrains.kotlin.plugin.serialization.gradle.plugin -> 1.9.0`
 - `org.jetbrains.kotlin.multiplatform:org.jetbrains.kotlin.multiplatform.gradle.plugin -> 1.9.0`